refactor tokenizer more
